move DbModule and DbGroup into independent files split from cluster.py
