+++ citrine/cluster.py
@@ -22,170 +22,13 @@ contents.
 #   Ideally, a cluster will be responsible for a single type of object so
 #   indexes and views can be efficiently written.
 import logging
-from os.path import join, exists
-from typing import Iterable
-from uuid import uuid4
+from os.path import join
+
+from ZODB import DB
 
 from citrine.citrinedb import CitrineDB
-from citrine.cluster_tools.dbmodule.dbmodule import create_dbmodule, \
-    delete_dbmodule, import_db, find_dbmodules
+from citrine.cluster_tools.dbmodule.dbgroup import DbGroup
 from citrine.cluster_tools.clusterdb.container import ClusterRegistryContainer
-
-
-class DbModule:
-    """
-    Manages the creation, migration, and deletion of a Citrine Database object
-    that is treated like an independent, importable code module.
-
-    This object is responsible for creating a directory structure with an
-    ``__init__.py`` file that provides a single importable object in the format
-    ``from <uuid> import get_db``
-    """
-
-    @property
-    def path(self):
-        """
-        Returns the path
-        :return:
-        """
-        return getattr(self, '___path___', None)
-
-    @path.setter
-    def path(self, value: str):
-        """
-        Sets the path. Only allows it to be set once and raises a
-        FileNotFoundError if attempting to set a path that does not exist
-        :param value: the path value
-        :return:
-        """
-        if hasattr(self, '___path___'):
-            raise AttributeError('Cannot change "path" attribute of DbModule')
-        if not exists(value):
-            raise FileNotFoundError(value)
-        setattr(self, '___path___', value)
-
-    def __init__(self, path: str = '.'):
-        self.path = path
-        self.db = import_db(self.path)
-        self.name = self.db.database_name
-
-    def open(self, transaction_manager=None, at=None, before=None):
-        """
-        Return a database Connection for use by application code.
-
-        Note that the connection pool is managed as a stack, to increase the
-        likelihood that the connection's stack will include useful objects.
-
-        :param transaction_manager: transaction manager to use, "None" will
-        default to a CitrineThreadTransactionManager
-        :param at: a ``datetime.datetime`` or 8 character transaction id of the
-        time to open the database with a read-only connection. Passing both
-        ``at`` and ``before`` raises a ValueError, and passing neither opens a
-        standard writable transaction of the newest state. A timezone-naive
-        ``datetime.datetime`` is treated as a UTC value.
-        :param before: like ``at``, but opens the readonly state before the tid
-        or datetime.
-        :return: CitrineConnection
-        """
-        return self.db.open(transaction_manager=transaction_manager,
-                            at=at, before=before)
-
-    @classmethod
-    def create(cls, path: str, name: str, overwrite: bool = False):
-        """
-        Creates a new module at the path location, if one does not exist.
-
-        The provided name will be given to the newly created database
-        """
-        module_path = create_dbmodule(path, name, overwrite)
-        return DbModule(module_path)
-
-    @classmethod
-    def destroy(cls, path: str, remove_empty: bool = True,
-                remove: bool = False):
-        """
-        Destroys a module at the path location
-        """
-        delete_dbmodule(path, remove_empty, remove)
-
-    def __call__(self):
-        return self.db
-
-
-class DbGroup:
-    """
-    A collection of databases as a single object. Will create databases as
-    importable modules under a common directory and provide managed storage
-    across all databases in the collection.
-
-    It is STRONGLY RECOMMENDED to use the discovery process, and ONLY the
-    discovery process, to manage the contents of the group, and is also strongly
-    encouraged to load an existing group rather than create a new one every
-    runtime.
-    """
-    @property
-    def databases(self):
-        """
-        Dictionary of names and database objects. Will derive a set from all
-        modules registered to the cluster unless a value has been manually set
-        """
-        if hasattr(self, '___databases___'):
-            return getattr(self, '___databases___', dict())
-        return {name: module.db for name, module in self.modules.items()}
-
-    @databases.setter
-    def databases(self, value):
-        setattr(self, '___databases___', value)
-
-    @databases.deleter
-    def databases(self):
-        delattr(self, '___databases___')
-
-    @property
-    def modules(self):
-        return getattr(self, '___modules___', dict())
-
-    @modules.setter
-    def modules(self, value):
-        setattr(self, '___modules___', value)
-
-    def __init__(self, root: str = '.', discovery=True, modules: dict = None):
-        """
-        Creates the pool, using the path as the root of the group. Any DbModules
-        in the root directory will be automatically retrieved and added to the
-        group if discovery is True
-        """
-        self.root = root
-        # uses modules if provided, discovers if instructed
-        self.modules = self.discover(root) | (modules if modules else dict()) \
-            if discovery else (modules if modules else dict())
-
-    def create_dbmodule(self, name: str = None, overwrite: bool = False):
-        """
-        Creates a new database using the provided name, at the path. If no
-        path is specified, the directory of execution will be used
-        :param name: The name to assign the database
-        """
-        new = DbModule.create(self.root, name=name, overwrite=overwrite)
-        self.modules[new.name] = new
-        return new
-
-    def destroy_dbmodule(self, name):
-        """
-        Destroys a db object from the pool entirely. This action is irreversible
-        :param name: The name of the database to be destroyed
-        """
-        DbModule.destroy(join(self.root, name))
-
-    @staticmethod
-    def discover(root):
-        """
-        Collects all database modules from the provided path and adds anything
-        into the pool that is not yet added
-        """
-        module_paths = find_dbmodules(root)
-        db_modules = [DbModule(path) for path in module_paths]
-        return {module.name: module for module in db_modules}
 
 
 class ClusterDB(DbGroup, CitrineDB):
@@ -246,9 +89,9 @@ class ClusterDB(DbGroup, CitrineDB):
             xrefs=xrefs, large_record_size=large_record_size, **storage_args)
         # zodb sets the databases value, so we reset to the derived version
         del self.databases
-        #self.validate_module_registration()
+        self.validate_module_registration()
 
-    def validate_module_registration(self):
+    def validate_module_registration(self, strict=False):
         """
         Checks that all detected modules match the previous registry, and logs
         inconsistencies if not.
@@ -263,12 +106,38 @@ class ClusterDB(DbGroup, CitrineDB):
         be accessible, but the contents will not be readable or written to by
         standard operations unless the database is explicitly instructed to
         merge the contents into the existing structure.
+
+        If strict is enabled and a check fails, an exception will be raised.
+
+        :param strict: whether to fail if a check fails
         """
+        found_modules = self.modules
         with self as conn:
-            known_modules = conn.container.modules
+            # check that we aren't missing modules and log modules we find that
+            # we aren't prepared to handle yet
+            known_modules = conn.container.modules.items()
             for name, module in known_modules.items():
-                if name not in self.modules.keys():
-                    self.logger.warning(f"Module {name} not found!")
+                if name not in found_modules.keys():
+                    raise ModuleNotFoundError(f'Expected module "{name}" not found')
+            for name, module in found_modules.items():
+                if name not in known_modules.keys():
+                    self.logger.info(f'Found unrecognized module "{name}"; ' +
+                                     'module will be held but UNUSED UNTIL REBUILD')
+            checkable_modules = {name: mod for name, mod in found_modules.items()
+                                 if name in known_modules.keys()}
+        # check that every found module has a database we can use
+        for name, module in checkable_modules.items():
+            db = module.db
+            # check the db is an instance of ZODB.DB, can be opened, and has
+            # CRUD functions (all of these are fatal errors)
+            if not isinstance(db, DB):
+                raise TypeError(f'db {name} is not a recognizable type of database!')
+            conn = db.open()
+            for name in ('create', 'read', 'update', 'delete'):
+                fn = getattr(conn, name, None)
+                if fn is None or not callable(fn):
+                    raise NotImplementedError(f'db missing necessary method {name}, ' +
+                                              'cannot proceed!')
 
 
     def create_dbmodule(self, name: str = None, overwrite: bool = False):
@@ -292,7 +161,6 @@ class ClusterDB(DbGroup, CitrineDB):
             with conn:
                 conn.delete(name)
                 DbGroup.destroy_dbmodule(self, name)
-
 
     @classmethod
     def new(cls, root, pool_size: int = 7,
